ESLint: update to v9
