--- .eslintignore
@@ -1,9 +1,0 @@
-.cache
-build
-build-module
-build-types
-node_modules
-packages/block-serialization-spec-parser/parser.js
-packages/react-native-editor/bundle
-vendor
-!.*.js
