+++ package.json
@@ -34,6 +34,8 @@
 		"@emotion/is-prop-valid": "1.2.2",
 		"@emotion/jest": "11.7.1",
 		"@emotion/native": "11.0.0",
+		"@eslint-community/eslint-plugin-eslint-comments": "4.4.0",
+		"@eslint/compat": "1.1.1",
 		"@geometricpanda/storybook-addon-badges": "2.0.5",
 		"@inquirer/prompts": "7.2.0",
 		"@octokit/rest": "16.26.0",
@@ -61,7 +63,7 @@
 		"@testing-library/react": "14.3.0",
 		"@testing-library/react-native": "12.4.3",
 		"@testing-library/user-event": "14.4.3",
-		"@types/eslint": "8.56.9",
+		"@types/eslint": "9.6.1",
 		"@types/estree": "1.0.5",
 		"@types/istanbul-lib-report": "3.0.0",
 		"@types/mime": "2.0.3",
@@ -101,16 +103,17 @@
 		"equivalent-key-map": "0.2.2",
 		"esbuild": "0.18.20",
 		"escape-html": "1.0.3",
+		"eslint": "9.11.0",
+		"eslint-config-prettier": "9.1.0",
 		"eslint-import-resolver-node": "0.3.4",
-		"eslint-plugin-eslint-comments": "3.1.2",
-		"eslint-plugin-import": "2.25.2",
-		"eslint-plugin-jest": "27.4.3",
-		"eslint-plugin-jest-dom": "5.0.2",
-		"eslint-plugin-prettier": "5.0.0",
+		"eslint-plugin-import": "2.3.0",
+		"eslint-plugin-jest": "28.8.3",
+		"eslint-plugin-jest-dom": "5.4.0",
+		"eslint-plugin-prettier": "5.2.1",
 		"eslint-plugin-react-compiler": "19.0.0-beta-0dec889-20241115",
-		"eslint-plugin-ssr-friendly": "1.0.6",
-		"eslint-plugin-storybook": "0.6.13",
-		"eslint-plugin-testing-library": "6.0.2",
+		"eslint-plugin-ssr-friendly": "1.3.0",
+		"eslint-plugin-storybook": "0.9.0--canary.156.ed236ca.0",
+		"eslint-plugin-testing-library": "6.3.0",
 		"execa": "4.0.2",
 		"fast-glob": "3.2.7",
 		"filenamify": "4.2.0",
