+++ packages/eslint-plugin/configs/custom.js
@@ -1,33 +1,35 @@
-module.exports = {
-	plugins: [ '@wordpress' ],
-	rules: {
-		'@wordpress/no-unused-vars-before-return': 'error',
-		'@wordpress/no-base-control-with-label-without-id': 'error',
-		'@wordpress/no-unguarded-get-range-at': 'error',
-		'@wordpress/no-global-active-element': 'error',
-		'@wordpress/no-global-get-selection': 'error',
-		'@wordpress/no-unsafe-wp-apis': 'error',
-		'@wordpress/no-wp-process-env': 'error',
+module.exports = [
+	{
+		plugins: {
+			'@wordpress': require('../plugin'),
+		},
+		rules: {
+			'@wordpress/no-unused-vars-before-return': 'error',
+			'@wordpress/no-base-control-with-label-without-id': 'error',
+			'@wordpress/no-unguarded-get-range-at': 'error',
+			'@wordpress/no-global-active-element': 'error',
+			'@wordpress/no-global-get-selection': 'error',
+			'@wordpress/no-unsafe-wp-apis': 'error',
+			'@wordpress/no-wp-process-env': 'error',
+		},
 	},
-	overrides: [
-		{
-			files: [ '*.native.js' ],
-			rules: {
-				'@wordpress/no-base-control-with-label-without-id': 'off',
-				'@wordpress/i18n-no-flanking-whitespace': 'error',
-				'@wordpress/i18n-hyphenated-range': 'error',
-			},
+	{
+		files: [ '*.native.js' ],
+		rules: {
+			'@wordpress/no-base-control-with-label-without-id': 'off',
+			'@wordpress/i18n-no-flanking-whitespace': 'error',
+			'@wordpress/i18n-hyphenated-range': 'error',
 		},
-		{
-			files: [
-				'*.test.js',
-				'**/test/*.js',
-				'packages/e2e-test-utils/**/*.js',
-			],
-			rules: {
-				'@wordpress/no-global-active-element': 'off',
-				'@wordpress/no-global-get-selection': 'off',
-			},
+	},
+	{
+		files: [
+			'*.test.js',
+			'**/test/*.js',
+			'packages/e2e-test-utils/**/*.js',
+		],
+		rules: {
+			'@wordpress/no-global-active-element': 'off',
+			'@wordpress/no-global-get-selection': 'off',
 		},
-	],
-};
+	},
+];
