+++ packages/eslint-plugin/configs/esnext.js
@@ -2,16 +2,16 @@
  * External dependencies
  */
 const { cosmiconfigSync } = require( 'cosmiconfig' );
+const ESLintParser = require( '@babel/eslint-parser' );
 
 const config = {
-	parser: '@babel/eslint-parser',
-	parserOptions: {
-		sourceType: 'module',
+	languageOptions: {
+		ecmaVersion: 6,
+		parser: ESLintParser,
+		parserOptions: {
+			sourceType: 'module',
+		},
 	},
-	env: {
-		es6: true,
-	},
-	extends: [ require.resolve( './es5.js' ) ],
 	rules: {
 		// Disable ES5-specific (extended from ES5)
 		'vars-on-top': 'off',
@@ -56,13 +56,13 @@ const config = {
 // https://github.com/davidtheclark/cosmiconfig/issues/246.
 const result = cosmiconfigSync( 'babel' ).search();
 if ( ! result || ! result.filepath ) {
-	config.parserOptions = {
-		...config.parserOptions,
-		requireConfigFile: false,
+	config.languageOptions.parserOptions = {
+		...config.languageOptions.parserOptions,
+		configFile: false,
 		babelOptions: {
 			presets: [ require.resolve( '@wordpress/babel-preset-default' ) ],
 		},
 	};
 }
 
-module.exports = config;
+module.exports = [ ...require( './es5.js' ), config ];
