+++ packages/eslint-plugin/configs/i18n.js
@@ -1,5 +1,7 @@
 module.exports = {
-	plugins: [ '@wordpress' ],
+	plugins: {
+		'@wordpress': require('../plugin'),
+	},
 	rules: {
 		'@wordpress/valid-sprintf': 'error',
 		'@wordpress/i18n-translator-comments': 'error',
