+++ packages/eslint-plugin/configs/test-e2e.js
@@ -1,11 +1,17 @@
+/**
+ * External dependencies
+ */
+const jest = require( 'eslint-plugin-jest' );
+const globals = require( 'globals' );
+
 module.exports = {
-	extends: [ 'plugin:jest/recommended' ],
-	env: {
-		browser: true,
-	},
-	globals: {
-		browser: 'readonly',
-		page: 'readonly',
-		wp: 'readonly',
+	...jest.configs[ 'flat/recommended' ],
+	languageOptions: {
+		globals: {
+			...globals.browser,
+			browser: 'readonly',
+			page: 'readonly',
+			wp: 'readonly',
+		},
 	},
 };
