+++ packages/eslint-plugin/configs/test-unit.js
@@ -1,9 +1,16 @@
-module.exports = {
-	extends: [ 'plugin:jest/recommended' ],
-	rules: {
-		'jest/expect-expect': [
-			'error',
-			{ assertFunctionNames: [ 'expect', 'measurePerformance' ] },
-		],
+/**
+ * External dependencies
+ */
+const jest = require( 'eslint-plugin-jest' );
+
+module.exports = [
+	jest.configs[ 'flat/recommended' ],
+	{
+		rules: {
+			'jest/expect-expect': [
+				'error',
+				{ assertFunctionNames: [ 'expect', 'measurePerformance' ] },
+			],
+		},
 	},
-};
+];
