+++ packages/eslint-plugin/index.js
@@ -1,4 +1,8 @@
-module.exports = {
-	configs: require( './configs' ),
-	rules: require( './rules' ),
-};
+/**
+ * Internal dependencies
+ */
+const plugin = require( './plugin' );
+
+Object.assign( plugin.configs, require( './configs' ) );
+
+module.exports = plugin;
