+++ packages/eslint-plugin/package.json
@@ -32,26 +32,25 @@
 	"main": "index.js",
 	"dependencies": {
 		"@babel/eslint-parser": "7.25.7",
-		"@typescript-eslint/eslint-plugin": "^6.4.1",
-		"@typescript-eslint/parser": "^6.4.1",
 		"@wordpress/babel-preset-default": "*",
 		"@wordpress/prettier-config": "*",
 		"cosmiconfig": "^7.0.0",
-		"eslint-config-prettier": "^8.3.0",
-		"eslint-plugin-import": "^2.25.2",
-		"eslint-plugin-jest": "^27.4.3",
-		"eslint-plugin-jsdoc": "^46.4.6",
-		"eslint-plugin-jsx-a11y": "^6.5.1",
-		"eslint-plugin-playwright": "^0.15.3",
-		"eslint-plugin-prettier": "^5.0.0",
-		"eslint-plugin-react": "^7.27.0",
-		"eslint-plugin-react-hooks": "^4.3.0",
+		"eslint-config-prettier": "^9.1.0",
+		"eslint-plugin-import": "^2.3.0",
+		"eslint-plugin-jest": "^28.8.3",
+		"eslint-plugin-jsdoc": "^50.2.4",
+		"eslint-plugin-jsx-a11y": "^6.10.0",
+		"eslint-plugin-playwright": "^1.6.2",
+		"eslint-plugin-prettier": "^5.2.1",
+		"eslint-plugin-react": "^7.36.1",
+		"eslint-plugin-react-hooks": "^5.1.0-rc-5d19e1c8-20240923",
 		"globals": "^13.12.0",
-		"requireindex": "^1.2.0"
+		"requireindex": "^1.2.0",
+		"typescript-eslint": "^8.6.0"
 	},
 	"peerDependencies": {
 		"@babel/core": ">=7",
-		"eslint": ">=8",
+		"eslint": ">=9",
 		"prettier": ">=3",
 		"typescript": ">=5"
 	},
