+++ packages/eslint-plugin/plugin.js
@@ -1,0 +1,13 @@
+/**
+ * Internal dependencies
+ */
+const { name, version } = require( './package.json' );
+
+module.exports = {
+	meta: {
+		name,
+		version,
+	},
+	configs: {},
+	rules: require( './rules' ),
+};
