--- packages/scripts/config/.eslintignore
@@ -1,3 +1,0 @@
-build
-node_modules
-vendor
