--- packages/scripts/config/.eslintrc.js
@@ -1,27 +1,0 @@
-/**
- * Internal dependencies
- */
-const { hasBabelConfig } = require( '../utils' );
-
-const eslintConfig = {
-	root: true,
-	extends: [ 'plugin:@wordpress/eslint-plugin/recommended' ],
-	overrides: [
-		{
-			// Unit test files and their helpers only.
-			files: [ '**/@(test|__tests__)/**/*.js', '**/?(*.)test.js' ],
-			extends: [ 'plugin:@wordpress/eslint-plugin/test-unit' ],
-		},
-	],
-};
-
-if ( ! hasBabelConfig() ) {
-	eslintConfig.parserOptions = {
-		requireConfigFile: false,
-		babelOptions: {
-			presets: [ require.resolve( '@wordpress/babel-preset-default' ) ],
-		},
-	};
-}
-
-module.exports = eslintConfig;
