+++ packages/scripts/package.json
@@ -58,7 +58,7 @@
 		"cssnano": "^6.0.1",
 		"cwd": "^0.10.0",
 		"dir-glob": "^3.0.1",
-		"eslint": "^8.3.0",
+		"eslint": "^9.11.0",
 		"expect-puppeteer": "^4.4.0",
 		"fast-glob": "^3.2.7",
 		"filenamify": "^4.2.0",
